# backend/.well-known/openapi.yaml
openapi: 3.0.0
info:
  title: InsightFlow MCP Server API
  description: MCP (Model Context Protocol) server for managing projects, stacks, roadmaps, tasks, and subtasks
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://your-domain.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase authentication token
paths:
...
  /mcp/projects/get:
    post:
      summary: Get project
      description: Get a single project by ID
      operationId: getProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
              properties:
                project_id:
                  type: string
      responses:
        '200':
          description: Project details
  /mcp/projects/create:
    post:
      summary: Create project
      description: Create a new project
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [active, completed, archived]
                database_schema:
                  type: object
      responses:
        '201':
          description: Project created
  /mcp/projects/update:
    post:
      summary: Update project
      description: Update an existing project
      operationId: updateProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
              properties:
                project_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                database_schema:
                  type: object
      responses:
        '200':
          description: Project updated
  /mcp/projects/delete:
    post:
      summary: Delete project
      description: Delete a project
      operationId: deleteProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
              properties:
                project_id:
                  type: string
      responses:
        '200':
          description: Project deleted
...
  /mcp/tasks/list:
    post:
      summary: List tasks
      description: Get a list of tasks for a roadmap
      operationId: listTasks
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
                page:
                  type: number
                limit:
                  type: number
      responses:
        '200':
          description: List of tasks
  /mcp/tasks/get:
    post:
      summary: Get task
      description: Get a single task with subtasks
      operationId: getTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
      responses:
        '200':
          description: Task details
  /mcp/tasks/create:
    post:
      summary: Create task
      description: Create a new task
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
                - name
              properties:
                roadmap_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [pending, in_progress, completed, blocked]
                priority:
                  type: string
                  enum: [low, medium, high]
                due_date:
                  type: string
      responses:
        '201':
          description: Task created
  /mcp/tasks/update:
    post:
      summary: Update task
      description: Update an existing task
      operationId: updateTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                priority:
                  type: string
                due_date:
                  type: string
      responses:
        '200':
          description: Task updated
  /mcp/tasks/delete:
    post:
      summary: Delete task
      description: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
      responses:
        '200':
          description: Task deleted
  /mcp/subtasks/list:
    post:
      summary: List subtasks
      description: Get a list of subtasks for a task
      operationId: listSubtasks
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
                page:
                  type: number
                limit:
                  type: number
      responses:
        '200':
          description: List of subtasks
  /mcp/subtasks/get:
    post:
      summary: Get subtask
      description: Get a single subtask
      operationId: getSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
      responses:
        '200':
          description: Subtask details
  /mcp/subtasks/create:
    post:
      summary: Create subtask
      description: Create a new subtask
      operationId: createSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
                - name
              properties:
                task_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: Subtask created
  /mcp/subtasks/update:
    post:
      summary: Update subtask
      description: Update an existing subtask
      operationId: updateSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Subtask updated
  /mcp/subtasks/delete:
    post:
      summary: Delete subtask
      description: Delete a subtask
      operationId: deleteSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
      responses:
        '200':
          description: Subtask deleted
...